requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==1.10.11
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-groq==0.2.1
langchain-pinecone==0.2.0
langchain_huggingface==0.1.2
langchain-text-splitters==0.3.2
unstructured==0.16.4
sentence-transformers==3.2.1
python-dotenv==1.0.1
pinecone==5.3.1
urllib3==1.26.15